+++ low/low.go
@@ -36,8 +36,8 @@ func DirectStop(pktsForFreeNumber int, buf []uintptr) {
 }
 
 // DirectSend sends one mbuf.
-func DirectSend(m *Mbuf, port uint8) bool {
-	return bool(C.directSend((*C.struct_rte_mbuf)(m), C.uint8_t(port)))
+func DirectSend(m *Mbuf, port uint16) bool {
+	return bool(C.directSend((*C.struct_rte_mbuf)(m), C.uint16_t(port)))
 }
 
 // Ring is a ring buffer for pointers
@@ -491,7 +491,7 @@ func FreeMempools() {
 	usedMempools = nil
 }
 
-func StopPort(port uint8) {
+func StopPort(port uint16) {
 	C.rte_eth_dev_stop(C.uint16_t(port))
 }
 
@@ -501,7 +501,7 @@ func GetPortsNumber() int {
 }
 
 // CreatePort initializes a new port using global settings and parameters.
-func CreatePort(port uint8, willReceive bool, sendQueuesNumber uint16, hwtxchecksum bool) error {
+func CreatePort(port uint16, willReceive bool, sendQueuesNumber uint16, hwtxchecksum bool) error {
 	addr := make([]byte, C.ETHER_ADDR_LEN)
 	var mempool *C.struct_rte_mempool
 	if willReceive {
@@ -510,7 +510,7 @@ func CreatePort(port uint8, willReceive bool, sendQueuesNumber uint16, hwtxcheck
 	} else {
 		mempool = nil
 	}
-	if C.port_init(C.uint8_t(port), C.bool(willReceive), C.uint16_t(sendQueuesNumber),
+	if C.port_init(C.uint16_t(port), C.bool(willReceive), C.uint16_t(sendQueuesNumber),
 		mempool, (*C.struct_ether_addr)(unsafe.Pointer(&(addr[0]))), C._Bool(hwtxchecksum)) != 0 {
 		msg := common.LogError(common.Initialization, "Cannot init port ", port, "!")
 		return common.WrapWithNFError(nil, msg, common.FailToInitPort)
